💉 Configured DI and fixed some Bugs

+++ src/backend/RateIdeas.Domain/Entities/Comments/CommentVote.cs
@@ -6,9 +6,9 @@ public class CommentVote : Auditable
 {
     public bool IsUpvote { get; set; }
 
-    public int UserId { get; set; }
+    public long UserId { get; set; }
     public User User { get; set; } = default!;
 
-    public int IdeaId { get; set; }
+    public long IdeaId { get; set; }
     public Comment Idea { get; set; } = default!;
 }
