+++ src/backend/RateIdeas.Domain/Entities/Ideas/IdeaVote.cs
@@ -6,9 +6,9 @@ public class IdeaVote : Auditable
 {
     public bool IsUpvote { get; set; }
 
-    public int UserId { get; set; }
+    public long UserId { get; set; }
     public User User { get; set; } = default!;
 
-    public int IdeaId { get; set; }
+    public long IdeaId { get; set; }
     public Idea Idea { get; set; } = default!;
 }
