+++ src/backend/RateIdeas.Domain/Entities/Ideas/SavedIdea.cs
@@ -4,9 +4,9 @@ namespace RateIdeas.Domain.Entities.Ideas;
 
 public class SavedIdea : Auditable
 {
-    public int UserId { get; set; }
+    public long UserId { get; set; }
     public User User { get; set; } = default!;
 
-    public int IdeaId { get; set; }
+    public long IdeaId { get; set; }
     public Idea Idea { get; set; } = default!;
 }
